requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
websocket-client
httpx
websockets
PyYAML
openai
